Modifying static text in review & OTP Confirmation. Moving Components files
